Refactor and fix tests for TaskRunner handler.
go unit tests should be in the same directory as the thing they are testing, so move email_test.go to the job package. It's actually testing the job handler's email dispatching functionality, so rename the file to be more clear. Then, refactor the method into actual go tests that run with go test.