Setup queue in case it is not yet created by publisher.
Proactively setup the RabbitMQ queue in case it is not already setup by
a publishing job. This allows the job to start cleanly (as opposed to
crashing when it can't consume from a nonexistent queue). Additionally,
this allows for easy debugging of the consumer side via manual message
publishing in RabbitMQ management interface.